Changes to MatchForm
- Copied over AbstractForm again (so it’s the most updated version)
- Reorganized declarations of visual element objects
- Added constants for FALSE & TRUE (for checkboxes)
- created StaticRecords array to hold records that only need to be
added once (unlike timestamp records)
- Created new Listener classes (that are more efficient + organized)
- Creation/changing of records in response to the visual elements —>
checkboxes, buttons, etc.
- for the teleop buttons —> the rawtimestamp records array list now
updates accurately to reflect a change in the edit texts. for example,
if the user increases the number in the edit text, then additional
records are added to the rawtimestamp records. if the user decreases
the number in the edit text (more likely), the last record is removed
until the # of records in the rawtimestamprecords w/ the same item id
is the same as the # in the edit text.
- attempt to get transfer working - not working yet
- save form functionality implemented